refactor(frontend): Separate module for token-group
