+++ src/frontend/src/lib/components/tokens/TokensDisplayHandler.svelte
@@ -3,7 +3,7 @@
 	import { combinedDerivedSortedNetworkTokensUi } from '$lib/derived/network-tokens.derived';
 	import { showZeroBalances } from '$lib/derived/settings.derived';
 	import type { TokenUi, TokenUiOrGroupUi } from '$lib/types/token';
-	import { groupTokensByTwin } from '$lib/utils/token.utils';
+	import { groupTokensByTwin } from '$lib/utils/token-group.utils';
 
 	// We start it as undefined to avoid showing an empty list before the first update.
 	export let tokens: TokenUiOrGroupUi[] | undefined = undefined;
